SAME Asymptotic Variance

Calculates the asymptotic variance (or variance - covariance matrix in the multidimensional case) of the SAME, given a specified family of distributions and the true parameter values.

Usage

avar_same(distr, ...)

Arguments

distr

A subclass of Distribution. The distribution family assumed.

...

extra arguments.

Value

A named matrix. The asymptotic covariance matrix of the estimator.

References

Ye, Z.-S. & Chen, N. (2017), Closed-form estimators for the gamma distribution derived from likelihood equations, The American Statistician 71(2), 177–181.

Van der Vaart, A. W. (2000), Asymptotic statistics, Vol. 3, Cambridge university press.

Tamae, H., Irie, K. & Kubokawa, T. (2020), A score-adjusted approach to closed-form estimators for the gamma and beta distributions, Japanese Journal of Statistics and Data Science 3, 543–561.

Mathal, A. & Moschopoulos, P. (1992), A form of multivariate gamma distribution, Annals of the Institute of Statistical Mathematics 44, 97–106.

Oikonomidis, I. & Trevezas, S. (2023), Moment-Type Estimators for the Dirichlet and the Multivariate Gamma Distributions, arXiv, https://arxiv.org/abs/2311.15025

See also

Examples

# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

# Simulation
set.seed(1)
shape1 <- 1
shape2 <- 2
D <- Beta(shape1, shape2)
x <- r(D)(100)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
#> [1] 24.27059
ll(x, c(shape1, shape2), D)
#> [1] 24.27059
ll(x, c(shape1, shape2), "beta")
#> [1] 24.27059

# Point Estimation - The e Functions

ebeta(x, type = "mle")
#>   shape1   shape2 
#> 1.063191 2.333189 
ebeta(x, type = "me")
#>    shape1    shape2 
#> 0.9988814 2.2451062 
ebeta(x, type = "same")
#>   shape1   shape2 
#> 1.041966 2.341944 

mle(x, D)
#>   shape1   shape2 
#> 1.063191 2.333189 
me(x, D)
#>    shape1    shape2 
#> 0.9988814 2.2451062 
same(x, D)
#>   shape1   shape2 
#> 1.041966 2.341944 

estim(x, D, type = "mle")
#>   shape1   shape2 
#> 1.063191 2.333189 

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
#>          shape1   shape2
#> shape1 1.597168 2.523104
#> shape2 2.523104 7.985838
vbeta(shape1, shape2, type = "me")
#>        shape1 shape2
#> shape1    2.1    3.3
#> shape2    3.3    9.3
vbeta(shape1, shape2, type = "same")
#>          shape1   shape2
#> shape1 1.644934 2.539868
#> shape2 2.539868 8.079736

avar_mle(D)
#>          shape1   shape2
#> shape1 1.597168 2.523104
#> shape2 2.523104 7.985838
avar_me(D)
#>        shape1 shape2
#> shape1    2.1    3.3
#> shape2    3.3    9.3
avar_same(D)
#>          shape1   shape2
#> shape1 1.644934 2.539868
#> shape2 2.539868 8.079736

avar(D, type = "mle")
#>          shape1   shape2
#> shape1 1.597168 2.523104
#> shape2 2.523104 7.985838