The d p q r Functions

Four generic functions that take a distribution object (e.g. Bern) and return the density, cumulative probability, quantile, and random generator functions, respectively.

Usage

d(x, ...)

p(x, ...)

qn(x, ...)

r(x, ...)

Arguments

x

an object of subclass Distribution.

...

extra arguments.

Value

The d p q r functions return the density, cumulative probability, quantile, and random generator functions, respectively.

Examples

# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

library(estim)

# Create the distribution
x <- Beta(3, 5)

# Density function
df <- d(x)
df(c(0.3, 0.8, 0.5))
#> [1] 2.268945 0.107520 1.640625

# Probability function
pf <- p(x)
pf(c(0.3, 0.8, 0.5))
#> [1] 0.3529305 0.9953280 0.7734375

# Density function
qf <- qn(x)
qf(c(0.3, 0.8, 0.5))
#> [1] 0.2763397 0.5167578 0.3641161

# Random Generator function
rf <- r(x)
rf(5)
#> [1] 0.5409149 0.5368579 0.4232217 0.3979307 0.1202451