Bernoulli Distribution

Bernoulli Distribution

Usage

Bern(prob = 0.5)

# S4 method for class 'Bern,numeric'
d(distr, x)

# S4 method for class 'Bern,numeric'
p(distr, x)

# S4 method for class 'Bern,numeric'
qn(distr, x)

# S4 method for class 'Bern,numeric'
r(distr, n)

# S4 method for class 'Bern'
mean(x)

# S4 method for class 'Bern'
median(x)

# S4 method for class 'Bern'
mode(x)

# S4 method for class 'Bern'
var(x)

# S4 method for class 'Bern'
sd(x)

# S4 method for class 'Bern'
skew(x)

# S4 method for class 'Bern'
kurt(x)

# S4 method for class 'Bern'
entro(x)

# S4 method for class 'Bern'
finf(x)

llbern(x, prob)

# S4 method for class 'Bern,numeric'
ll(distr, x)

ebern(x, type = "mle", ...)

# S4 method for class 'Bern,numeric'
mle(distr, x)

# S4 method for class 'Bern,numeric'
me(distr, x)

vbern(prob, type = "mle")

# S4 method for class 'Bern'
avar_mle(distr)

# S4 method for class 'Bern'
avar_me(distr)

Arguments

prob

numeric. The distribution parameter.

distr

an object of class Bern.

x

an object of class Bern. If the function also has a distr argument, x is a numeric vector, a sample of observations.

n

numeric. The sample size.

type

character, case ignored. The estimator type (mle, me, or same).

...

extra arguments.

Value

Each type of function returns a different type of object:

Distribution Functions: When supplied with one argument (distr), the d(), p(), q(), r(), ll() functions return the density, cumulative probability, quantile, random sample generator, and log-likelihood functions, respectively. When supplied with both arguments (distr and x), they evaluate the aforementioned functions directly.

Moments: Returns a numeric, either vector or matrix depending on the moment and the distribution. The moments() function returns a list with all the available methods.

Estimation: Returns a list. The estimator of the unknown parameters. Note that in distribution families like the binomial, multinomial, and negative binomial, the size is not returned, since it is considered known.

Variance: Returns a named matrix. The asymptotic covariance matrix of the estimator.

See also